create docu p1 mk placeholder

+++ gendeb.sh
@@ -2,7 +2,7 @@
 # at the moment only generate dir structure in /build
 
 echo "clear build"
-rm -v -R ./build
+[ -d ./build ] && rm -v -R ./build
 mkdir build
 
 echo "copy files"
